each Layer gets its own canvas, enabling Layer.show/hide
- zIndex is still specified as before, but the now zIndex is bound to
each canvas with .data(). Then as each Layer is added, the DOM order is
sorted with d3.selection.sort.   New Layers default to zIndex 1, the
base canvas is 0, and the HUD is Number.MAX_VALUE.
- HUD is on by default, though won’t draw the status bar unless
configured.  It makes more sense for the hud to contain the graticule
since it sits on top.
- pushed some methods down to Layer and Grid